app/requirements.txt
fastapi==0.115.12
joblib==1.4.2
keras==3.9.2
matplotlib==3.10.1
numpy==2.2.4
opencv_python==4.11.0.86
Pillow==11.2.1
pydicom==3.0.1
scikit_learn==1.6.1
torch==2.6.0
torchvision==0.21.0
